+++ src/Resources/contao/config/config.php
@@ -13,7 +13,6 @@ $GLOBALS['BE_MOD']['content']['news']['tables'][] = 'tl_news_list_archive';
  */
 $GLOBALS['TL_MODELS']['tl_news_list']         = '\HeimrichHannot\NewsBundle\Model\NewsListModel';
 $GLOBALS['TL_MODELS']['tl_news_list_archive'] = '\HeimrichHannot\NewsBundle\Model\NewsListArchiveModel';
-$GLOBALS['TL_MODELS']['tl_news_tags']         = '\HeimrichHannot\NewsBundle\Model\NewsTagsModel';
 
 /**
  * Hooks