+++ src/Resources/contao/languages/en/tl_news.php
@@ -3,13 +3,6 @@
 $arrLang = &$GLOBALS['TL_LANG']['tl_news'];
 
 /**
- * Fields
- */
-$arrLang['tags']                    = ['Tags', 'Assign one or more tags.'];
-
-
-/**
  * Legends
  */
-$arrLang['tags_legend']           = 'Tags';
 $arrLang['readers_survey_legend'] = 'Readers Survey';